+++ src/utilities/theme/utils.ts
@@ -17,9 +17,26 @@ export function buildCustomProperties(
     : buildLegacyColors(themeConfig);
 }
 
-export function buildThemeContext(themeConfig: ThemeConfig): Theme {
+export function buildThemeContext(
+  themeConfig: ThemeConfig,
+  cssCustomProperties?: CustomPropertiesLike,
+): Theme {
   const {logo} = themeConfig;
-  return {logo};
+  return {
+    logo,
+    // eslint-disable-next-line babel/camelcase
+    UNSTABLE_cssCustomProperties: toString(cssCustomProperties),
+  };
+}
+
+function toString(obj?: CustomPropertiesLike) {
+  if (obj) {
+    return Object.entries(obj)
+      .map((pair) => pair.join(':'))
+      .join(';');
+  } else {
+    return undefined;
+  }
 }
 
 /* eslint-disable babel/camelcase */